// src/common/enums/error-code.enum.ts
export enum ErrorCode {
    AUTH_HEADER_MISSING = 'AUTH_HEADER_MISSING',
    INVALID_BEARER_TOKEN = 'INVALID_BEARER_TOKEN',
    USER_NOT_FOUND_OR_UNVERIFIED = 'USER_NOT_FOUND_OR_UNVERIFIED',
    FORBIDDEN_RESOURCE = 'FORBIDDEN_RESOURCE',
    ORG_NOT_FOUND = 'ORG_NOT_FOUND',
    MEMBER_NOT_FOUND = 'MEMBER_NOT_FOUND',
    USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
    SLUG_CONFLICT = 'SLUG_CONFLICT',
    SLUG_ALREADY_EXISTS = 'SLUG_ALREADY_EXISTS',
    INVALID_EMAIL = 'INVALID_EMAIL',
    EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
    TOO_MANY_ATTEMPTS = 'TOO_MANY_ATTEMPTS',
    TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
    INVALID_PASSWORD = 'INVALID_PASSWORD',
    USER_NOT_FOUND = 'USER_NOT_FOUND',
    USER_ALREADY_VERIFIED = 'USER_ALREADY_VERIFIED',
    REFRESH_TOKEN_MISSING = 'REFRESH_TOKEN_MISSING',
    CUSTOMER_NOT_FOUND = 'CUSTOMER_NOT_FOUND',
    ADDRESS_NOT_FOUND = 'ADDRESS_NOT_FOUND',
    EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
    CUSTOMER_EMAIL_ALREADY_EXISTS = 'CUSTOMER_EMAIL_ALREADY_EXISTS',
    TOKEN_EXPIRED = 'TOKEN_EXPIRED',
    INVALID_TOKEN = 'INVALID_TOKEN',
    INVALID_TOKEN_TYPE = 'INVALID_TOKEN_TYPE',
    ORG_MEMBER_NOT_FOUND = 'ORG_MEMBER_NOT_FOUND',
    OWNER_CANNOT_BE_MODIFIED = 'OWNER_CANNOT_BE_MODIFIED',
    CANNOT_DELETE_SELF = 'CANNOT_DELETE_SELF',
    CANNOT_DELETE_OWNER = 'CANNOT_DELETE_OWNER',
    USER_ID_REQUIRED = 'USER_ID_REQUIRED',
    UNKNOWN = 'UNKNOWN',
    MISSING_CUSTOMER_ID = 'MISSING_CUSTOMER_ID',
    SERVICE_REQ_NOT_FOUND = 'SERVICE_REQ_NOT_FOUND',
    SKILL_NOT_FOUND = 'SKILL_NOT_FOUND',
    SKILL_ASSOCIATED = 'SKILL_ASSOCIATED',
    TECHNICIAN_ALREADY_EXISTS = 'TECHNICIAN_ALREADY_EXISTS',
    INVALID_EMPLOYEE_ROLE = 'INVALID_EMPLOYEE_ROLE',
    TECHNICIAN_NOT_FOUND = 'TECHNICIAN_NOT_FOUND',
    NOT_ALLOWED = 'NOT_ALLOWED',
    SERVICE_NOT_FOUND = 'SERVICE_NOT_FOUND',
    SERVICE_NOT_DELETABLE = 'SERVICE_NOT_DELETABLE',
    SERVICE_NOT_UPDATABLE = 'SERVICE_NOT_UPDATABLE',
    AVAILABILITY_NOT_FOUND = 'AVAILABILITY_NOT_FOUND',
    INVALID_AVAILABILITY_TIME = 'INVALID_AVAILABILITY_TIME',
    AVAILABILITY_OVERLAP = 'AVAILABILITY_OVERLAP',
    SCHEDULE_NOT_FOUND = 'SCHEDULE_NOT_FOUND',
    SCHEDULE_CONFLICT = 'SCHEDULE_CONFLICT',
}